% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoCorrelationDendrogram.R
\name{autoCorrelationDendrogram}
\alias{autoCorrelationDendrogram}
\title{Reduces multicollinearity automatically through repeated application of \code{\link{correlationDendrogram}}.}
\usage{
autoCorrelationDendrogram(
  x,
  variables = NULL,
  exclude.variables = c("x", "y", "presence"),
  correlation.threshold = 0.5,
  biserialCorrelation.output = NULL,
  plot = TRUE,
  label.size = 6
  )
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis. Defaults to \code{c("x", "y", "presence")}.}

\item{correlation.threshold}{Numeric in the interval [0, 1], maximum Pearson correlation of the selected variables. Defaults to 0.5.}

\item{biserialCorrelation.output}{List, output of the function \code{\link{biserialCorrelation}}. Its R-squared scores are used to select variables.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}

\item{label.size}{Numeric, size of the dendrogram labels.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Computes the correlation between all pairs of variables in a training dataset and removes variables until there are no more variables correlated above a given threshold (0.5 Pearson correlation by default). This function requires a \code{\link{biserialCorrelation}} output, and it won't run if the argument remains \code{NULL}.
}
\examples{
\dontrun{
data("virtualSpeciesPB")

bis.cor <- biserialCorrelation(
 x = virtualSpeciesPB,
 exclude.variables = c("x", "y")
)

selected.vars <- autoCorrelationDendrogram(
 x = virtualSpeciesPB,
 variables = NULL,
 exclude.variables = c("x", "y", "presence"),
 correlation.threshold = 0.5,
 biserialCorrelation.output = bis.cor
)$selected.variables
}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
