% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSelectVariables.R
\name{autoSelectVariables}
\alias{autoSelectVariables}
\title{Automatic selection of predictive variables for species distribution modeling}
\usage{
autoSelectVariables(
  x,
  presence.column = "presence",
  variables = NULL,
  exclude.variables = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{presence.column}{Character, name of the presence column.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis.}

\item{verbose}{Boolean, defaults to TRUE. Triggers messages describing what variables are being removed.}
}
\value{
A character vector with the names of the selected variables.

#' @examples
\dontrun{
data("virtualSpeciesPB")
selected.vars <- autoSelectVariables(
 x = virtualSpeciesPB,
 presence.column = "presence",
 variables = NULL,
 exclude.variables = c("x", "y"),
 verbose = TRUE
)
selected.vars
}
}
\description{
Computes the biserial correlation between presence and background data for a set of predictors with \code{\link{biserialCorrelationPB}}, and uses the output of this function as argument \code{try.to.keep} in \code{\link{autoVIF}}. This approach should provide a reasonable outcome when modelling with presence-background data. If you have a preference on what variables to use in your modelling exercise, please use \code{\link{autoVIF}} instead.
}
\author{
Blas Benito <blasbenito@gmail.com>.
}
