% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoVIF.R
\name{autoVIF}
\alias{autoVIF}
\title{Automatic variable selection with variance inflation factor (VIF) analysis}
\usage{
autoVIF(x)
}
\arguments{
\item{x}{A data frame with numeric columns.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Selects variables within a dataframe that are not correlated with each other, or with linear combinations of other variables, by using the variance inflation factor (VIF) criteria implemented in the \code{\link[HH]{vif}} function.
}
\author{
Blas Benito <blasbenito@gmail.com>
}
