% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biserialCorrelation.R
\name{biserialCorrelation}
\alias{biserialCorrelation}
\title{Biserial correlation analysis of presence and background data for variable selection}
\usage{
biserialCorrelation(
  x,
  presence.column = "presence",
  variables = NULL,
  exclude.variables = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{presence.column}{Character, name of the presence column.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}
}
\value{
A named list with two slots named \code{plot} and \code{df}. The former contains a ggplot object with the biserial correlation analysis. The latter is a data frame with the following columns:
\itemize{
  \emph{variable}: Name of the predictive variable.
  \emph{R2}: R-squared of the biserial correlation.
  \emph{p}: p-value of the correlation analysis.
}
The output data frame is ordered, starting with the higher R2 values.
}
\description{
Computes the biserial correlation between presence and background data for a set of predictors. A high biserial correlation for a given predictor indicates that the distributions of the presence and background records are separated enough in the space of predictor values to suggest that the predictor is a good candidate for a species distribution model.
}
\examples{
data(virtualSpeciesPB)
cPB <- biserialCorrelation(
  x = virtualSpeciesPB,
  presence.column = "presence",
  variables = c("bio1", "bio5", "bio6")
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
