% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationDendrogram.R
\name{correlationDendrogram}
\alias{correlationDendrogram}
\title{Correlation dendrogram to help reduce multicollinearity in a training dataset.}
\usage{
correlationDendrogram(
  x,
  variables = NULL,
  exclude.variables = NULL,
  correlation.threshold = 0.5,
  automatic.selection = TRUE,
  biserialCorrelation.output = NULL,
  plot = TRUE
  )
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis.}

\item{correlation.threshold}{Numeric in the interval [0, 1], maximum Pearson correlation of the selected variables.}

\item{automatic.selection}{Boolean. If \code{TRUE}, the function provides a vector of selected variables along with the dendrogram plot. Otherwise, only the dendrogram plot is returned.}

\item{biserialCorrelation.output}{List, output of the function \code{\link{biserialCorrelation}}. Its R-squared scores are used to select variables.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}
}
\value{
If \code{automatic.selection = TRUE}, a list with two slots named "dendrogram" (a ggplot2 object) and "selected.variables" with the dendrogram and the character vector with the selected variables. Otherwise, only returns the dendrogram.

#' @examples
\dontrun{
data("virtualSpeciesPB")

bis.cor <- biserialCorrelation(
 x = virtualSpeciesPB,
 exclude.variables = c("x", "y")
)

selected.vars <- correlationDendrogram(
 x = virtualSpeciesPB,
 variables = NULL,
 exclude.variables = c("x", "y", "presence"),
 correlation.threshold = 0.5,
 automatic.selection = TRUE,
 biserialCorrelation.output = bis.cor
)$selected.variables
}
}
\description{
Computes the correlation between all pairs of variables in a training dataset. If a \code{\link{biserialCorrelation}} output is provided, it further selects variables automatically based on the R-squared value obtained by each variable in the biserial correlation analysis.
}
\author{
Blas Benito <blasbenito@gmail.com>.
}
