% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfFrom4D.R
\name{dfFrom4D}
\alias{dfFrom4D}
\title{Transforms 4D object into a data frame}
\usage{
dfFrom4D(x)
}
\arguments{
\item{x}{Named list, 4D object produced by \code{\link{import4D}}.}
}
\value{
A data frame with all the valid cells of the raster bricks contained in \code{x}. The slot names of \code{x} are written in the column "time".
}
\description{
Generates a data frame from a 4D object produced by \code{\link{import4D}}.
}
\author{
Blas Benito <blasbenito@gmail.com>
}
