% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMworkshop.R
\docType{data}
\name{europe21kBP}
\alias{europe21kBP}
\title{Predictive variables for Europe representing climatic conditions during the Last Glacial Maximum (21 ka BP).}
\format{Raster brick with 19 layers.}
\usage{
library(raster)
data(europe21kBP)
plot(europe21kBP)
}
\description{
Raster brick with predictive variables to fit species distribution models for the European territory at ~20km resolution and reference system WGS84 (units are degrees of latitude and longitude, EPSG is 4326).
The variables are:
\itemize{
  \item \emph{bio1}: annual mean temperature (ºC x 10).
  \item \emph{bio2}: mean diurnal temperature range, computed as \eqn{mean of monthly (max temp - min temp)}.
  \item \emph{bio3}: isothermality, computed as \eqn{(bio2/bio7)*100}.
  \item \emph{bio4}: temperature seasonality (standard deviation of monthly averages * 100).
  \item \emph{bio5}: maximum temperature of the warmest month.
  \item \emph{bio6}: minimum temperature of the coldest month.
  \item \emph{bio7}: temperature annual range, computed as \eqn{(bio5-bio6}.
  \item \emph{bio8}: mean temperature of the wettest quarter.
  \item \emph{bio9}: mean temperature of the driest quarter.
  \item \emph{bio10}: mean temperature of the warmest quarter.
  \item \emph{bio11}: mean temperature of the coldest quarter.
  \item \emph{bio12}: annual precipitation (mm.).
  \item \emph{bio13}: precipitation of wettest month.
  \item \emph{bio14}: precipitation of driest month.
  \item \emph{bio15}: precipitation seasonality (standard deviation of monthly averages * 100).
  \item \emph{bio16}: precipitation of wettest quarter.
  \item \emph{bio17}: precipitation of driest quarter.
  \item \emph{bio18}: precipitation of warmest quarter.
  \item \emph{bio19}: precipitation of coldest quarter.
}
}
\details{
Bioclim variables (\emph{bioXX}) were downloaded from the CCSM4 model available in the \emph{Last Glacial Maximum section} of the \emph{Worldclim v1.4} dataset (Hijmans et al. 2005, \url{https://www.worldclim.org/version1}).
}
\references{
Hijmans, R.J., S.E. Cameron, J.L. Parra, P.G. Jones and A. Jarvis, 2005. Very high resolution interpolated climate surfaces for global land areas. International Journal of Climatology 25: 1965-1978.
}
\keyword{datasets}
