% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVirtualSpecies.R
\name{makeVirtualSpecies}
\alias{makeVirtualSpecies}
\title{Generates virtual species}
\usage{
makeVirtualSpecies(
variables,
niche.parameters = NULL,
seed = NULL,
species.type = "multiplicative",
max.n = NULL
)
}
\arguments{
\item{variables}{A raster brick or stack with three or more environmental variables.}

\item{niche.parameters}{A named list containing the parameters of the gaussian functions representing the niche of the species. Each slot name must be a layer name of the brick/stack \code{variables}. The slot must contain a vector with two numbers. The first represents the mean of a normal distributión, and the second represents the standard deviation of the same normal distribution. These numbers must be in the same units of the \code{variables} layer with the same name of the slot. If this argument is set to \code{NULL}, then a random set of niche parameters is generated by the function based on an uncorrelated set of predictive variables.}

\item{seed}{A integer determining a random seed. Only relevant when \code{niche.parameters == NULL}. Added to allow reproducibility in the generation of random virtual species.}

\item{species.type}{String, one of "additive" or "multiplicative". This is an argument of the function \code{\link[virtualspecies]{generateSpFromFun}}, which is used internally. It defines how the overall environmental suitability is computed from the response functions of each predictor. These are summed when \code{species.type = "additive"}, and multiplied when \code{species.type = "multiplicative"}, which is the default behavior.}

\item{max.n}{Numeric integer, maximum number of presence records to extract if possible, otherwise the function returns as many presence records as possible.}
}
\value{
A named list with 5 slots:
\itemize{
 \item \emph{niche.dimensions}: Names of the variables used to define the niche of the virtual species.
 \item \emph{niche.parameters}: Either the same \code{niche.parameters} used as input, or the niche definition generated randomly by the function if \code{niche.parameters = NULL}.
 \item \emph{niche.plot}: A ggplot with the density plots of the presence records (represented in blue) and the background (represented in yellow), and a habitat suitability map with presence records.
 \item \emph{suitability.raster}: A habitat suitability raster generated from the niche parameters by the function \code{\link[virtualspecies]{generateSpFromFun}}.
 \item \emph{observed.presence}: A dataframe with the coordinates x and y (in the same coordinate reference system as \code{variables}) of the selected presences.
}
}
\description{
Generates virtual species from virtual ecological niches. The niche definition can be either provided by the user, or generated randomly.
}
\details{
This function relies on the functions \code{\link[virtualspecies]{generateSpFromFun}}, \code{\link[virtualspecies]{convertToPA}} ,\code{\link[virtualspecies]{sampleOccurrences}} of the \emph{virtualspecies} package, so please, cite Leroy et al. (2016) if you find this function useful.

Note that for simplicity, the settings of the functions mentioned above have been simplified. Particularly, the function \code{\link[virtualspecies]{convertToPA}} is configured with the parameters \code{PA.method = "probability"}, \code{prob.method = "linear"}, \code{a = 1}, and \code{b = 0}. This setting is the safest one when \code{niche.parameters} is not provided, and therefore randomized virtual taxa are being generated.

The function \code{\link[virtualspecies]{sampleOccurrences}} is configured as follows: \code{type = "presence only"}; \code{correct.by.suitability = TRUE}.

Please, check the help files of these three functions to better understand how they work!
}
\references{
Leroy, B., Meynard, C.N., Bellard, C. and Courchamp, F. (2016), virtualspecies, an R package to generate virtual species distributions. Ecography, 39: 599-607. doi:10.1111/ecog.01388
}
\author{
Blas Benito <blasbenito@gmail.com>
}
