% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRaster.R
\name{plotRaster}
\alias{plotRaster}
\title{Pretty plots of a raster file}
\usage{
plotRaster(x, n = 100, opacity = 0.5, begin = 0, end = 1,
  direction = 1, option = "D")
}
\arguments{
\item{x}{A raster object. It can be the subset of a brick, as in \code{x[["variable_name"]]}.}

\item{n}{Number of colors (integer higher than 1) to use in the color palette. Argument of the \code{\link[viridis]{viridis}} function.}

\item{opacity}{Transparency in the range [0, 1]. If close to 1, the underlying political map will remain hidden.}

\item{begin}{Numeric in the range [0, 1], color at which the color palette starts.}

\item{end}{Numeric in the range [0, 1], color at which the color palette ends}

\item{direction}{Order of the colors in the color palette. Use -1 to reverse the color palette.}

\item{option}{A character string indicating the colormap option to use. Four options are available: "magma" (or "A"), "inferno" (or "B"), "plasma" (or "C"), "viridis" (or "D", the default option) and "cividis" (or "E").}
}
\value{
A leaflet plot.
}
\description{
Uses the \code{\link[leaflet]} library to plot a raster file.

plotRaster(
x,
n = 100,
opacity = 0.5,
begin = 0,
end = 1,
direction = 1,
option = "D"
)
}
\author{
Blas Benito <blasbenito@gmail.com>
}
