% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUseAvailability.R
\name{plotUseAvailability}
\alias{plotUseAvailability}
\title{Plots density functions of use versus availability from a training data frame}
\usage{
plotUseAvailability(
  x,
  presence.column = "presence",
  variables = NULL,
  exclude.variables = NULL,
  plot = TRUE
)
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{presence.column}{Character, name of the presence column.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis.}

\item{plot}{Boolean, prints biserial correlation plot if \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
Plots the relative densities of ones (presence, a.k.a, "use") versus zeroes (background, a.k.a, "availability") for each environmental predictor in a training data frame. This plot helps to understand the relationship between use and availability in order to make informed decisions during variable selection. When for a given variable the density of use peaks over low availability it indicates that the species selects those values of a variable at a higher rate than what is expected by chance. On the other hand, variables with a very high overlap between use and availability will likely turn out to have a low predictive value during SDM fitting.
}
\examples{
data("virtualSpeciesPB")
x <- plotUseAvailability(
 x = virtualSpeciesPB,
 presence.column = "presence",
 variables = NULL,
 exclude.variables = c("x", "y"),
 plot = TRUE
)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
