% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareTrainingData.R
\name{prepareTrainingData}
\alias{prepareTrainingData}
\title{Prepares training data for SDMs}
\usage{
prepareTrainingData(
  xy,
  variables,
  n = 10000,
  presence.only = FALSE,
  background = TRUE,
  restricted.background = FALSE,
  restricted.background.buffer = 200,
  plot = TRUE,
  thinning = FALSE,
  minimum.distance = raster::xres(variables)
  )
}
\arguments{
\item{xy}{A data frame with two columns with coordinates x and y representing presence records. Column names are irrelevant as long as the first column represents the x coordinate, and the second column represents the y coordinate.}

\item{variables}{A raster brick or stack with environmental variables. Must be in the same reference system of \code{xy}.}

\item{n}{Integer, number of background points to generate. To generate pseudo-absences just define a low n (i.e. twice the number of presences).}

\item{presence.only}{Boolean. If \code{TRUE}, all other options are set to \code{FALSE}, and the function returns presence-only data.}

\item{background}{Boolean. If \code{TRUE}, all other options are set to \code{FALSE}, and the function returns background data generated to the extension of \code{variables}.}

\item{restricted.background}{Boolean. If \code{TRUE}, all other options are set to \code{FALSE}, and the function returns restricted background data generated within a buffer of width equeal to \code{restricted.background.buffer} around the presence records \code{xy}.}

\item{restricted.background.buffer}{Integer, buffer in kilometres around \code{xy} over which to define the buffer delimiting the restricted background.}

\item{plot}{Boolean. If \code{TRUE}, the output data is plotted.}

\item{thinning}{Boolean. If \code{TRUE}, the function \code{\link{reduceSpatialCorrelation}} is applied to \code{xy} to reduce the spatial aggregation of the data.}

\item{minimum.distance}{Numeric, minimum distance between consecutive points in the output dataset. Defaults to the resolution of \code{variables}. The minimum distance can be extracted from the resolution of \code{variables}, as in  \code{minimum.distance <- xres(variables)}.}
}
\value{
A data frame ready to fit an species distribution model.
}
\description{
Produces "presence-only", "background" and "restricted background" data to fit species distribution models. The background is selected at random within the provided raster brick or stack, while the restricted background is selected within a buffer (usually based on the maximum dispersal distance of the target species) around the presence records. The selection of background points is made by the function \code{\link[dismo]{randomPoints}}, from the \code{dismo} package (Hijmans et al. 2017). The function can also apply thinning to the presence coordinates to reduce spatial autocorrelation through the function \code{\link{reduceSpatialCorrelation}}. To produce pseudo-absences instead of background data, just reduce the argument \code{n} to a number between the number and presences and twice the number of presences.
}
\examples{
\dontrun{
data(virtualSpecies)
data(europe2000)

#presence-only data
presence.only <- prepareTrainingData(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  n,
  presence.only = TRUE,
  plot = TRUE
)

#background
background <- prepareTrainingData(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  n,
  background = TRUE,
  plot = TRUE
)

#restricted background
restricted.background <- prepareTrainingData(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  n,
  restricted.background = TRUE,
  restricted.background.buffer = 100,
  plot = TRUE
)

#applying thinning
restricted.background <- prepareTrainingData(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  n = 1000,
  restricted.background = TRUE,
  restricted.background.buffer = 100,
  plot = TRUE,
  thinning = TRUE,
  minimum.distance = raster::xres(europe2000)
)
}

}
\references{
Robert J. Hijmans, Steven Phillips, John Leathwick and Jane Elith (2017). dismo: Species Distribution Modeling. R package version 1.1-4. https://CRAN.R-project.org/package=dismo
}
\author{
Blas Benito <blasbenito@gmail.com>. The function \code{\link[dismo]{randomPoints}} is authored by Robert J. Hijmans.
}
