% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMworkshop.R
\docType{data}
\name{presenceBackground}
\alias{presenceBackground}
\alias{virtualSpecies}
\title{Presence and background data for a virtual species}
\format{Data frame.}
\usage{
data(presenceBackground)
names(presenceBackground)
}
\description{
Dataframe with 8344 observations and 34 variables containing presence records of a virtual species and background data. Each record contains the latitude, longitude, and values of 31 predictive variables, along with a presence column with 1 indicating presence, and 0 indicating background. This data is ready for model fitting.
}
\keyword{datasets}
