% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceSpatialCorrelation.R
\name{reduceSpatialCorrelation}
\alias{reduceSpatialCorrelation}
\title{Applies "thinning" to a set of coordinates to reduce their spatial correlation}
\usage{
reduceSpatialCorrelation(
xy,
variables,
minimum.distance = NULL,
random.start = FALSE,
seed = NULL,
verbose = FALSE
)
}
\arguments{
\item{xy}{A data frame with two columns with coordinates x and y. Column names are irrelevant as long as the first column represents the x coordinate, and the second column represents the y coordinate.}

\item{variables}{A raster brick or stack with environmental variables. Must be in the same reference system of \code{xy}.}

\item{minimum.distance}{Numeric, minimum distance between consecutive points in the output dataset. Defaults to the resolution of \code{variables}. The minimum distance can be extracted from the resolution of \code{variables}, as in  \code{min.dist <- xres(variables)}.}

\item{random.start}{Boolean, defaults to \code{FALSE}). If \code{TRUE}, the \code{xy} dataset is randomly reshuffled so the start of the thinning changes, and with that, the output dataset.}

\item{seed}{Integer determining a random seed. Only relevant when \code{random.start = TRUE}. Added to allow reproducibility in the generation of datasets with a random start.}

\item{verbose}{Boolean. If \code{FALSE} (default), all messages are supressed.}
}
\value{
A data frame with the same columns as \code{xy}, but a lower number of records.
}
\description{
This function reduces the spatial clustering of a set of presence records. It is intended to reduce spatial autocorrelation, and reduce sampling bias, particularly at larger geographical scales. It takes as input a set of coordinates, a brick or stack of environmental variables, and a minimum distance, and returns a new set of coordinates in which the distance between adjacent points is equal or higher than the established minimum distance. This operation is named "thinning", and helps to reduce the spatial correlation of a presence dataset. This function applies thinning, but preserves the presence records representing the extremes of any of the predictive variables provided.
}
\examples{
data("virtualSpecies")
data(europe2000)
xy.thinned <- reduceSpatialCorrelation(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  minimum.distance = 4
)
xy.thinned

#generating datasets with different starting points
#generates a different dataset with different nrow on each run
xy.thinned <- reduceSpatialCorrelation(
  xy = virtualSpecies$observed.presence,
  variables = europe2000,
  minimum.distance = 4,
)
nrow(virtualSpecies$observed.presence)
nrow(xy.thinned)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
