% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMworkshop.R
\docType{package}
\name{SDMworkshop-package}
\alias{SDMworkshop}
\alias{SDMworkshop-package}
\title{SDMworkshop: Streamlining the Development of Species Distribution Models}
\description{
Helps developing species distribution models (SDM) with functions to easily plot data, select variables, assess multicollinearity, etc etc
}
\details{
Details
}
