% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMworkshop.R
\docType{data}
\name{virtualSpecies}
\alias{virtualSpecies}
\title{Named list describing a virtual species.}
\format{Named list.}
\usage{
virtualSpecies
}
\description{
Named list with 5 named slots containing a description of a virtual species generated with the function \code{\link{makeVirtualSpecies}}, which depends on the function \code{\link[virtualspecies]{generateSpFromFun}} (Leroy et al. 2015).
The list slots are:
\itemize{
  \item \emph{niche.dimensions}: vector with the names of the variables in \code{\link{europe2000}} used to define the ecological niche of the virtual species.
  \item \emph{niche.parameters}: data frame with the mean (column \emph{mean}) and the standard deviation (column \emph{sd}) of the normal functions used to define the ecological niche of the virtual species.
  \item \emph{niche.plot}: plot describing the niche and distribution of the species.
  \item \emph{suitability}: raster map, suitability raster for the virtual species as defined by \code{\link[virtualspecies]{generateSpFromFun}}.
  \item \emph{observed.presence}: data frame with the coordinates \emph{x} and \emph{y} (reference system WGS84, EPSG 4326) of the simulated presences of the virtual species.
}
}
\references{
Leroy B, Meynard CN, Bellard C, Courchamp F (2015). “virtualspecies, an R package to generate virtual species distributions.” Ecography. doi: 10.1111/ecog.01388.
}
\keyword{datasets}
