% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightPB.R
\name{weightPB}
\alias{weightPB}
\title{Computes weights for presence and background records}
\usage{
weightPB(x)
}
\arguments{
\item{x}{Ordered numeric vector with ones and zeroes respectively representing presence and background records. This will generally be the "presence" column of a data frame to be used to fit a species distribution model.}
}
\value{
A numeric vector with the weights of the ones and zeroes available in the input vector.
}
\description{
Presence and background samples are highly unbalanced, generally by more than one order of magnitude. Many functions such as \code{\link[mgcv]{gam}} or \code{\link[ranger]{ranger}} have an argument named \code{weights} or similar that helps to work around the sample unbalance. Such an argument requires a numeric vector as input, with one position per record in the data used to fit the model. Each value in the weights vector represents the relative weight of a single case, and all the values in the weights vector get to sum 1. For example, if a presence column is \code{1, 0, 0}, its relative weights would be \code{0.5, 0.25, 0.25}. The function takes as input an ordered numeric vector of ones and zeroes representing the presence and background data of an input dataset to fit SDMs.
}
\examples{
data(virtualSpeciesPB)
weights <- weightPB(x = virtualSpeciesPB$presence)
table(weights)

}
\author{
Blas Benito <blasbenito@gmail.com>
}
