% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoSelectVariables.R
\name{autoSelectVariables}
\alias{autoSelectVariables}
\title{Automatic selection of predictive variables for species distribution modeling}
\usage{
autoSelectVariables(
  x,
  presence.column = "presence",
  variables = NULL,
  exclude.variables = NULL,
  plot = TRUE,
  label.size = 6
)
}
\arguments{
\item{x}{A data frame with a presence column with 1 indicating presence and 0 indicating background, and columns with predictor values.}

\item{presence.column}{Character, name of the presence column.}

\item{variables}{Character vector, names of the columns representing predictors. If \code{NULL}, all numeric variables but \code{presence.column} are considered.}

\item{exclude.variables}{Character vector, variables to exclude from the analysis.}

\item{plot}{Boolean, if \code{TRUE}, prints last correlation dendrogram to test the final output.}

\item{label.size}{Numeric, size of the dendrogram labels.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Applies \code{\link{biserialCorrelation}}, \code{\link{correlationDendrogram}} (with correlation threshold set to 0.5), and \code{\link{autoVIF}} to automatically select a set of non-correlated variables with the higher biserial correlation as possible.
}
\examples{
\dontrun{
data(virtualSpeciesPB)
selected.vars <- autoSelectVariables(
  x = virtualSpeciesPB,
  presence.column = "presence",
  exclude.variables = c("x", "y")
)
selected.vars
HH::vif(virtualSpeciesPB[, selected.vars])
cor(virtualSpeciesPB[, selected.vars])
}

}
\author{
Blas Benito <blasbenito@gmail.com>.
}
