% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoVIF.R
\name{autoVIF}
\alias{autoVIF}
\title{Automatic variable selection with variance inflation factor (VIF) analysis}
\usage{
autoVIF(
  x,
  try.to.keep = NULL,
  verbose = TRUE
)
}
\arguments{
\item{x}{A data frame with numeric columns.}

\item{try.to.keep}{A character vector with the names of the variables the user would like to keep, in order of preference. If this argument is not \code{NULL}, the function first applies \code{\link[HH]{vif}} to the variables not in \code{x} that are not in \code{try.to.keep}, then to the variables in \code{try.to.keep}, and finally to the outcome of both vif analyses, while always trying to remove variables not in \code{try.to.keep}. It is recommended to use the variable order of the \code{variable} column from the output of \code{\link{biserialCorrelation}}.}

\item{verbose}{Boolean, defaults to TRUE. Triggers messages describing what variables are being removed.}
}
\value{
A character vector with the names of the selected variables.
}
\description{
Selects variables within a dataframe that are not correlated with each other, or with linear combinations of other variables, by using the variance inflation factor (VIF) criteria implemented in the \code{\link[HH]{vif}} function (Heilberger and Holland 2004).
}
\examples{
\dontrun{
data("europe2000")
df <- raster::as.data.frame(europe2000[[c("bio1", "bio5", "bio6", "bio11", "bio12")]])
selected.vars <- SDMworkshop::autoVIF(
 x = df,
 try.to.keep = c("bio5", "bio6", "bio1"),
 verbose = TRUE
)
selected.vars

#autoVIF can also take the output of SDMworkshop::biserialCorrelation
#as try.to.keep argument, as follows:
data(virtualSpeciesPB)

cPB <- SDMworkshop::biserialCorrelation(
x = virtualSpeciesPB,
presence.column = "presence",
variables = c("bio1", "bio5", "bio6")
)

#note that cPB$df$variable is ordered from
#higher to lower biserial correlation
#higher biserial correlation is linked
#to higher predictive importance
selected.vars <- SDMworkshop::autoVIF(
 x = df,
 try.to.keep = cPB$df$variable,
 verbose = TRUE
)
selected.vars

}

}
\references{
Heiberger, Richard M. and Holland, Burt (2004). Statistical Analysis and Data Display: An Intermediate Course with Examples in S-Plus, R, and SAS. Springer Texts in Statistics. Springer. ISBN 0-387-40270-5.
}
\author{
Blas Benito <blasbenito@gmail.com>. The function \code{\link[HH]{vif}} is authored by Richard M. Heiberger <rmh@temple.edu>.
}
