% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDMworkshop.R
\docType{package}
\name{SDMworkshop-package}
\alias{SDMworkshop}
\alias{SDMworkshop-package}
\title{SDMworkshop: Streamlining the Development of Species Distribution Models}
\description{
Helps developing species distribution models with functions to easily plot data, select variables, evaluate spatial correlation, apply thinning to presence data. Furthermore, it provides functions to fit and evaluate different models.
}
\details{
Details
}
